+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- added `resolveValue` func
+
+### Changed
+
+- `editAt` & `editPropAt` now check if given index are in range - if not, it will return the original array reference
+- `editAt` now accept updater func as value too
+- `editPropAt` now use editAt with updater func inside
+
 ## [2.6.0] - 2025-10-22
 
 ### Added